+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Toolium Changelog
 v2.1.1
 ------
 
-*Release date: In development*
+*Release date: 2021-09-22*
 
 - Avoid to overwrite parent in group elements when a custom parent is defined
 - Fix Chrome options to allow to configure them at the same time in *Chrome* sections and in *goog:chromeOptions*